Always use the W3C Trace Context Propagator for extracting context in OpenTelemetryInterceptorFactory
