+++ sdk-interceptor-opentelemetry/src/main/java/dev/restate/sdk/interceptor/opentelemetry/OpenTelemetryAttributes.java
@@ -8,9 +8,16 @@
 // https://github.com/restatedev/sdk-java/blob/main/LICENSE
 package dev.restate.sdk.interceptor.opentelemetry;
 
+import dev.restate.sdk.endpoint.HeadersAccessor;
 import io.opentelemetry.api.common.AttributeKey;
+import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
+import io.opentelemetry.context.propagation.TextMapGetter;
+import io.opentelemetry.context.propagation.TextMapPropagator;
+import org.jspecify.annotations.NonNull;
+import org.jspecify.annotations.Nullable;
 
 /** Standard OpenTelemetry attribute keys used by the Restate interceptors. */
+@org.jetbrains.annotations.ApiStatus.Internal
 public final class OpenTelemetryAttributes {
   private OpenTelemetryAttributes() {}
 
@@ -21,4 +28,20 @@ public final class OpenTelemetryAttributes {
   public static final AttributeKey<String> INVOCATION_TARGET =
       AttributeKey.stringKey("restate.invocation.target");
   public static final AttributeKey<String> RUN_NAME = AttributeKey.stringKey("restate.run.name");
+
+  public static final TextMapGetter<HeadersAccessor> HEADERS_GETTER =
+      new TextMapGetter<>() {
+        @Override
+        public Iterable<String> keys(HeadersAccessor carrier) {
+          return carrier.keys();
+        }
+
+        @Nullable
+        @Override
+        public String get(@Nullable HeadersAccessor carrier, @NonNull String key) {
+          return carrier == null ? null : carrier.get(key);
+        }
+      };
+  public static final TextMapPropagator W3C_TRACE_CONTEXT_PROPAGATOR =
+      W3CTraceContextPropagator.getInstance();
 }
